+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - format message lines starting with `>` as quotes #3434
 - node: remove `split2` dependency #3418
 - node: add git installation info to readme #3418
+- limit the rate of webxdc update sending #3417
 
 ### Fixes
 - set a default error if NDN does not provide an error